+++ CHANGELOG.md
@@ -9,7 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-### Deleted
+### Removed
+
+## [v0.1.1](https://github.com/isu-avista/portal/releases/tag/v0.1.1) - 2021-02-15
+### Added
+* Added the basic data route to allow for data to enter the portal from the IoT devices
+* Added the docker-compose file to run the system
+* Added systemd service file
+* Added install script
+
+### Changed
+* Updated existing tests
+* Modified the existing server to correspond to changes in base-server
+* Modified the dockerfile to build the docker image
+* Updated README to reflect new instructions for installing and running
+
+### Removed
 
 ## [v0.1.0](https://github.com/isu-avista/portal/releases/tag/v0.1.0) - 2020-12-15
 ### Added
@@ -20,4 +35,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-### Deleted
+### Removed